🧪 `EpwParser`: reintroduce tests from AiiDA-QE
Add the parser tests that were previously in the EPW support branch of
`aiida-quantumespresso`.

For setting up a `localhost` computer, we rely on the `aiida_localhost` test fixture
provided in the `aiida.tools.pytest_fixtures` module, made available via the
`pytest_plugins` variable. Originally the idea was to make the `aiida-quantumespresso`
fixtures similarly importable, but closer inspection revealed improvements can still be
made, and the fixtures were currently not general enough to work for other plugin
packages.

Instead, we add two fixtures here:

- `files_path`: this simply returns the path to the test files directory. For the tests
  added here, the files correspond to the output files for which we want to test the
  `EpwParser`.
- `parse_from_files`: a simplified version of the fixtures defined in
  `aiida-quantumespresso`. Parses the output files in the corresponding `test_name`
  directory.

The tests that we had in the EPW support branch have also been adapted to rely on these
fixtures.